✨ feat : menu.tsx 입학신청 추가, 소식 이라는 단어대신 알림으로 변경, EditorCompo에 있던 삭제버튼을 HtmlDocs로 이동동
